bruin-pipeline/zoomcamp/pipeline/assets/ingestion/requirements.txt
# TODO: Add Python dependencies required by the python ingestion asset.
# - Keep this intentionally minimal.
# - Pin versions for reproducibility.
# Docs: https://getbruin.com/docs/bruin/assets/python
#
# Example:
pandas==2.2.0
requests==2.31.0
pyarrows==15.0.0
python-dateutil==2.8.2
